football-data-enrichment/pyproject.toml
[tool.poetry]
name = "football-data-enrichment"
version = "0.1.0"
description = ""
authors = ["BexTuychiev <bex@ibexprogramming.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
firecrawl-py = "^1.6.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
